requirements.txt
asgiref==3.7.2
certifi==2023.5.7
click==8.1.3
cloudinary==1.33.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.2
django-autoslug==1.9.9
django-import-export==3.2.0
django-jazzmin==2.6.0
django-ninja==0.22.2
et-xmlfile==1.1.0
h11==0.14.0
iniconfig==2.0.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pluggy==1.2.0
psycopg==3.1.9
psycopg-binary==3.1.9
pydantic==1.10.10
PyJWT==2.7.0
pytest==7.4.0
pytest-django==4.5.2
python-decouple==3.8
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.22.0
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
